added market priceing and wierd cirular error

+++ prediction-core/src/main/java/pl/frej/waw/prediction/core/boundary/UserController.java
@@ -1,5 +1,5 @@
 package pl.frej.waw.prediction.core.boundary;
 
 public interface UserController {
-    Integer getFunds(String id);
+    Long getFunds(String id);
 }
